projects/JobHunter/requirements.txt
streamlit
pandas
pdfminer.six
spacy
plotly
phonenumbers
feedparser
beautifulsoup4
requests
